feat(app): Add FlatList
