+++ src/screens/Home.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { View, Text, TouchableOpacity } from 'react-native';
+
+const Home = ({ navigation }) => {
+  return (
+    <View>
+      <TouchableOpacity onPress={() => navigation.navigate('ColorPalette')}>
+        <Text>Solarized</Text>
+      </TouchableOpacity>
+    </View>
+  );
+};
+
+export default Home;
